Move methods to Optimizer/Driver
