+++ dawn/src/dawn/Compiler/Driver.h
@@ -20,47 +20,9 @@
 #include "dawn/IIR/StencilInstantiation.h"
 #include "dawn/Optimizer/Driver.h"
 #include "dawn/Optimizer/Options.h"
-#include "dawn/Serialization/IIRSerializer.h"
 #include "dawn/Serialization/SIRSerializer.h"
 
-#include <list>
-#include <map>
-#include <memory>
-#include <string>
-
 namespace dawn {
-
-/// TODO Move these to Optimizer/Driver when OptimizerContext is removed
-/// {
-/// @brief Lower to IIR and run groups
-std::map<std::string, std::shared_ptr<iir::StencilInstantiation>>
-run(const std::shared_ptr<SIR>& stencilIR, const std::list<PassGroup>& groups,
-    const Options& options = {});
-
-/// @brief Lower to IIR and run groups. Use strings in place of C++ structures.
-///
-/// NOTE: This method always returns the stencil instantiations as json string objects, not
-/// bytes, as this greatly simplifies the conversion.
-/// See https://pybind11.readthedocs.io/en/stable/advanced/cast/strings.html for more details.
-std::map<std::string, std::string> run(const std::string& sir, dawn::SIRSerializer::Format format,
-                                       const std::list<dawn::PassGroup>& groups,
-                                       const dawn::Options& options = {});
-
-/// @brief Run groups
-std::map<std::string, std::shared_ptr<iir::StencilInstantiation>>
-run(const std::map<std::string, std::shared_ptr<iir::StencilInstantiation>>&
-        stencilInstantiationMap,
-    const std::list<PassGroup>& groups, const Options& options = {});
-
-/// @brief Run groups. Use strings in place of C++ structures.
-///
-/// NOTE: This method always returns the stencil instantiations as json string objects, not bytes,
-/// as this greatly simplifies the conversion.
-/// See https://pybind11.readthedocs.io/en/stable/advanced/cast/strings.html for more details.
-std::map<std::string, std::string>
-run(const std::map<std::string, std::string>& stencilInstantiationMap, IIRSerializer::Format format,
-    const std::list<PassGroup>& groups, const Options& options = {});
-/// }
 
 /// @brief Convenience function to compile SIR directly to a translation unit
 std::unique_ptr<codegen::TranslationUnit> compile(
